Add support for HEVC 10-bit downsampling into NV12
If userspace requests NV12M 8-bit but the bitstream is 10-bit, we take
advantage of the decoder's "double write" feature where it decodes the
frames using Framebuffer Compression, but outputs a second set of 8-bit
NV12 buffers.

This special mode however requires twice the usual amount of memory.
For 4K, with 16 CAPTURE buffers and 2 OUTPUT buffers, you're looking at
~450MB of ram usage.